+++ examples/binning.py
@@ -1,11 +1,13 @@
+import asyncio
 from datetime import timedelta
 
+from tau.core import NetworkScheduler
 from tau.event import Do
 from tau.signal import Interval, BufferWithCount, BufferWithTime
 
-from tau.testing import TestSchedulerContextManager
 
-with TestSchedulerContextManager(shutdown_delay=30) as scheduler:
+async def main():
+    scheduler = NetworkScheduler()
     network = scheduler.get_network()
     values = Interval(scheduler)
     Do(network, values, lambda: print(f"input values: {values.get_value()}"))
@@ -13,8 +15,9 @@ with TestSchedulerContextManager(shutdown_delay=30) as scheduler:
     buffer1 = BufferWithCount(network, values, count=2)
     Do(network, buffer1, lambda: print(f"buffer1 values: {buffer1.get_value()}"))
 
-    buffer2 = BufferWithTime(network, values, timedelta(seconds=5), scheduler=scheduler.get_native_scheduler())
+    buffer2 = BufferWithTime(network, values, timedelta(seconds=5))
     Do(network, buffer2, lambda: print(f"buffer2 values: {buffer2.get_value()}"))
 
-
+asyncio.get_event_loop().create_task(main())
+asyncio.get_event_loop().run_forever()