+++ examples/running_sum.py
@@ -1,11 +1,16 @@
-from tau.event import Do
+import asyncio
+
+from tau.core import NetworkScheduler
+from tau.event import Lambda
 from tau.math import RunningSum
 from tau.signal import From
 
-from tau.testing import TestSchedulerContextManager
 
-with TestSchedulerContextManager() as scheduler:
+async def main():
+    scheduler = NetworkScheduler()
     network = scheduler.get_network()
     values = From(scheduler, [0.0, 3.2, 2.1, 2.9, 8.3, 5.7])
     total = RunningSum(network, values)
-    Do(network, total, lambda x: print(f'{x:.2f}'))
+    Lambda(network, total, lambda x: print(f'{x[0].get_value():.2f}'))
+
+asyncio.run(main())