+++ HISTORY.rst
@@ -3,6 +3,13 @@
 Release History
 ---------------
 
+0.1.0 (2020-04-04)
++++++++++++++++++++
+
+- Remove dependency on APScheduler
+- Rewrite to use asyncio internally
+- Added websocket example
+
 0.0.2 (2020-03-28)
 +++++++++++++++++++